chore(rename-files): Rename files
This commit renames component files to match their class names

[Delivers #158750005]
