+++ src/components/auth/LogoutUser.js
@@ -6,7 +6,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import '../../static/css/style.css';
 import axios from "axios";
 import { apiUrl } from '../../App';
-import { isLoggedIn, clearUser } from '../../utils/helpers';
+import { isLoggedIn, clearUser } from '../../utils/Helpers';
 
 class LogoutUser extends Component {
 	constructor() {
