+++ src/components/auth/ProfileUser.js
@@ -4,7 +4,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import '../../static/css/style.css';
 import axios from "axios";
 import { apiUrl } from '../../App';
-import { isLoggedIn, clearUser } from '../../utils/helpers';
+import { isLoggedIn, clearUser } from '../../utils/Helpers';
 
 class ProfileUser extends Component {
 	constructor(props) {
