+++ src/components/auth/RegisterUser.js
@@ -7,7 +7,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import '../../static/css/style.css';
 import axios from "axios";
 import { apiUrl } from '../../App'
-import { contains } from '../../utils/helpers';
+import { contains } from '../../utils/Helpers';
 
 class RegisterUser extends Component {
 	constructor() {
