+++ src/components/businesses/ReviewBusiness.js
@@ -6,7 +6,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import '../../static/css/style.css';
 import axios from "axios";
 import { apiUrl } from '../../App'
-import { isLoggedIn } from '../../utils/helpers';
+import { isLoggedIn } from '../../utils/Helpers';
 
 class ReviewBusiness extends Component {
 	constructor(props) {
