+++ src/components/businesses/UpdateBusiness.js
@@ -5,7 +5,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import '../../static/css/style.css';
 import axios from "axios";
 import { apiUrl } from '../../App';
-import { isLoggedIn } from '../../utils/helpers';
+import { isLoggedIn } from '../../utils/Helpers';
 
 class UpdateBusiness extends Component {
 	constructor(props) {
