+++ src/tests/auth/RegisterUser.test.js
